Updated browsertime to newest version (22.6.0)
This used to be 14.8.0, so quite the upgrade.

I am editing this commit to undo my removal of the `chromeDevtoolsProtocol.js` and `har.js` files. These have some relationship with the "perfLog" in chromium, and generate some needed files in the testbed. Both of these files live in `/usr/src/app/lib/chrome/` in the normal `sitespeedio/browsertime` container. It looks like the original developers of this project copy these modified versions over to get chromium to use them. Since these files changes significantly between 14 and 22, I compared the diff of 14 with the modified version, and tried to apply the same changes starting with the 22 version.

also added --no-cache to the build file