Added mitmproxy container to capture sites
This docker container can be used to generate .save files in the
mahimahi format. it runs a server on port 8080, so the browser/client
must be pointed to it.

coming soon will be an implementation that utilizes the browsertime
container to generate these, so we can best cache what it will request
during the testbed run