updated certificates
these were set to expire in 2023, so I regenerated them. these are used
to copy into the browsertime container when the image is built (in
build.sh), and in setup/certificates/cert.sh where the h2o.py and
quiche.py scripts call it to create their respective certs which are
copied after the container starts.

it'd probably be easier to copy the certs over to the browsertime
container after it starts too...