made improvements to the browsertime container
namely, I removed the build step that copied the chrome(/driver)
binaries, which drastically increases the build-time for the container.
instead, the binaries are accessed via a mounted docker volume that maps
to the `chromium/src/out/Default` directory. this means that if chromium
ever needs to be rebuilt, the container does not need to be.